requirement.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==4.2.2
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
heroku==0.1.4
idna==3.4
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==1.5
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0